Clear, delete, and operator buttons update screen
Clear button clears screen.
Delete button deletes last entry on screen.
Operator buttons are added to screen when clicked. They don't appear if there are no numbers currently on the screen. Additionally, when the user attempts adding multiple operator symbols in a row, the last operator is instead replaced with the new one. This does not currently work when user attempts an equation more than 3 separate numbers long.
Current js is pretty ugly; lots of nested if statements. Hoping to fix this later.