Calculator checks when operator is present
Updated containsOperator function and created lastCharIsOperation function.
containsOperator now returns an array; if string does not contain operator, array is [undefined, undefined], and if string does contain operator, array is [true, (operator present)]. This is to prepare for a function that will take the equation off the calculator screen and solve it.
lastCharIsOperator is self-explanatory. Uses containsOperator to assess the last character in a string and returns true or false.